+++ src/components/CalendarMonth.jsx
@@ -21,7 +21,7 @@ import toISODateString from '../utils/toISODateString';
 import ModifiersShape from '../shapes/ModifiersShape';
 import ScrollableOrientationShape from '../shapes/ScrollableOrientationShape';
 import DayOfWeekShape from '../shapes/DayOfWeekShape';
-import BaseClass, { pureComponentAvailable } from '../utils/baseClass';
+
 
 import {
   HORIZONTAL_ORIENTATION,
@@ -90,8 +90,7 @@ const defaultProps = {
   verticalBorderSpacing: undefined,
 };
 
-/** @extends React.Component */
-class CalendarMonth extends BaseClass {
+class CalendarMonth extends React.PureComponent {
   constructor(props) {
     super(props);
 
@@ -274,4 +273,4 @@ export default withStyles(({ reactDates: { color, font, spacing } }) => ({
     paddingTop: 12,
     paddingBottom: 7,
   },
-}), { pureComponent: pureComponentAvailable })(CalendarMonth);
+}), { pureComponent: typeof React.PureComponent !== 'undefined' })(CalendarMonth);