+++ stories/DayPicker.js
@@ -1,5 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
+import { withInfo } from '@storybook/addon-info';
+import DirectionProvider, { DIRECTIONS } from 'react-with-direction/dist/DirectionProvider';
 import DayPicker from '../src/components/DayPicker';
 
 import {
@@ -46,25 +48,25 @@ const TestCustomInfoPanel = () => (
 );
 
 storiesOf('DayPicker', module)
-  .addWithInfo('default', () => (
+  .add('default', withInfo()(() => (
     <DayPicker />
-  ))
-  .addWithInfo('with custom day size', () => (
+  )))
+  .add('with custom day size', withInfo()(() => (
     <DayPicker daySize={50} />
-  ))
-  .addWithInfo('single month', () => (
+  )))
+  .add('single month', withInfo()(() => (
     <DayPicker numberOfMonths={1} />
-  ))
-  .addWithInfo('3 months', () => (
+  )))
+  .add('3 months', withInfo()(() => (
     <DayPicker numberOfMonths={3} />
-  ))
-  .addWithInfo('vertical', () => (
+  )))
+  .add('vertical', withInfo()(() => (
     <DayPicker
       numberOfMonths={2}
       orientation={VERTICAL_ORIENTATION}
     />
-  ))
-  .addWithInfo('vertically scrollable with 12 months', () => (
+  )))
+  .add('vertically scrollable with 12 months', withInfo()(() => (
     <div
       style={{
         height: 568,
@@ -76,57 +78,81 @@ storiesOf('DayPicker', module)
         orientation={VERTICAL_SCROLLABLE}
       />
     </div>
-  ))
-  .addWithInfo('vertical with custom day size', () => (
+  )))
+  .add('vertical with custom day size', withInfo()(() => (
     <DayPicker
       numberOfMonths={2}
       orientation={VERTICAL_ORIENTATION}
       daySize={50}
     />
-  ))
-  .addWithInfo('vertical with custom height', () => (
+  )))
+  .add('vertical with custom height', withInfo()(() => (
     <DayPicker
       numberOfMonths={2}
       orientation={VERTICAL_ORIENTATION}
       verticalHeight={568}
     />
-  ))
-  .addWithInfo('with custom arrows', () => (
+  )))
+  .add('vertical with DirectionProvider', withInfo()(() => (
+    <DirectionProvider direction={DIRECTIONS.RTL}>
+      <DayPicker
+        numberOfMonths={2}
+        orientation={VERTICAL_ORIENTATION}
+        isRTL
+      />
+    </DirectionProvider>
+  )))
+  .add('vertically scrollable with DirectionProvider', withInfo()(() => (
+    <DirectionProvider direction={DIRECTIONS.RTL}>
+      <div
+        style={{
+          height: 568,
+          width: 320,
+        }}
+      >
+        <DayPicker
+          numberOfMonths={12}
+          orientation={VERTICAL_SCROLLABLE}
+        />
+      </div>
+    </DirectionProvider>
+  )))
+  .add('with custom arrows', withInfo()(() => (
     <DayPicker
       navPrev={<TestPrevIcon />}
       navNext={<TestNextIcon />}
     />
-  ))
-  .addWithInfo('with custom details', () => (
+  )))
+  .add('with custom details', withInfo()(() => (
     <DayPicker
       renderDayContents={day => (day.day() % 6 === 5 ? '😻' : day.format('D'))}
     />
-  ))
-  .addWithInfo('vertical with fixed-width container', () => (
+  )))
+  .add('vertical with fixed-width container', withInfo()(() => (
     <div style={{ width: '400px' }}>
       <DayPicker
         numberOfMonths={2}
         orientation={VERTICAL_ORIENTATION}
       />
     </div>
-  ))
-  .addWithInfo('with info panel', () => (
+  )))
+  .add('with info panel', withInfo()(() => (
     <DayPicker
       renderCalendarInfo={() => (
         <TestCustomInfoPanel />
       )}
     />
-  ))
-  .addWithInfo('with custom week day format', () => (
+  )))
+  .add('with custom week day format', withInfo()(() => (
     <DayPicker
       weekDayFormat="ddd"
     />
-  ))
-  .addWithInfo('with no animation', () => (
+  )))
+  .add('with no animation', withInfo()(() => (
     <DayPicker
       transitionDuration={0}
     />
-  ))
-  .addWithInfo('noBorder', () => (
+  )))
+  .add('noBorder', withInfo()(() => (
     <DayPicker noBorder />
-  ));
+  )));